Good-enough starting point
I removed all the existing perldoc from FastLook.pm :x
need to look into those more since I need to learn to be
more concise (and much more selective about what I want
to explain...)

I also didn't add the test files because they were
originally meant for List::BisectSearch, which was like
a fraction of this interface (with left search only,
no search-any variants either); the NEW stuff with
syntactic sugar and arybase (Array::Base) compatibility
and everything definitely needs to be tested too,
which -- agh, i'm going to re-organize as well....

I'm not sure if this is too much comments for non-POD
stuff.  I tried my best to make every comment fit on
one screen so it's not conceptually too overwhelming.
I see the point of separating user-level docs, which
enjoys more abstraction and less nitty-gritty details;
but at the same time, I find it difficult to read the
code itself due to how clean it is in contrast!
I hope this has the effect of bridging the gap between
the two layers, so everyone has something to look at
when descending the hierarchy of abstraction.