Create SECURITY.md
... with help from the GitHub web interface.

(actually forget about it ugh, GitHub adds ugly spaces and
takes newlines from places they should be in, and throws
them in places where they shouldn't be.  AND it insists on
committing with the name I use in https://github.com/rapidcow
(which I do not want to put my real name in).  I genuinely
do not want to put up with this horrendous tasteless GitHub
web interface anymore (ㆆ_ㆆ) NO signature for you !!!)