The (x) => (1 - x) was likely not a very good map
I was trying to avoid going into negative numbers, thinking it
would be nice to not have to go over the negatives (if we had
unsigned integers)... but in increasing I've only bumped up the
[-S, 1) range to [1 - S, 2)!

Frankly speaking it would have made more sense if we were
looking at the un-mirrored side (-1, S], and raise that so it's
(0, S + 1] and it is completely lifted off the negative.  But
having to map THAT is completely beyond what I was trying to
achieve with this example....  There's nothing about negative
numbers to be afraid of when we're at such a high-level place,
so just use the easiest map that doesn't confuse me. :)

Also, I guess this promises that there will be multiple parts
to this... I'm thinking three parts, with the last one finally
leaning away from the math and towards the actual interface,
which does mean that by the time I get there, I might have a
good idea of what I might want to put in the Tutorial/Cookbook.
So -- not a total shameless plug for math! ;P